feat: rename vote script files
