+++ scripts/vote_2025_05_07_hoodi_dg_launch.py
