+++ tests/test_2025_05_07_hoodi_dg_launch.py
@@ -4,7 +4,7 @@ import os
 from typing import NamedTuple, List
 from brownie import web3, chain, interface, ZERO_ADDRESS, accounts, reverts
 from hexbytes import HexBytes
-from scripts.vote_05_05_2025_hoodi_dg_launch import start_vote
+from scripts.vote_2025_05_07_hoodi_dg_launch import start_vote
 from utils.config import contracts
 from brownie.network.transaction import TransactionReceipt
 from utils.test.tx_tracing_helpers import *
